requirements.txt
anthropic
chromadb
pymupdf
fastapi
uvicorn
python-dotenv
pydantic
sentence-transformers
